Make iterator member private in goto_trace_step
This commit aims to strengthen the invariants on the `pc` iterator
member of goto_trace_stept objects. By restricting access through
setters and getters, we can record when a 'legitimate' value is assigned
to `pc`. Then, we can check this flag, rather than comparing the `pc`
field to a default-constructed iterator, which causes undefined
behaviour.